Translate the "plays" collection
There are 3 main things I did here:

1. Added a _data/translations.yml file

For on-page text, link text, alt text, the site name, etc, it seems
like the convention is to create a yml file with "en" and "fr" keys
and swap between them based on the current language.

Gets pretty verbose in the template but it's simple and easy.

2. Added translations for the plays files

It's a bit weird what we're doing with collections here, it seems
like most of the time, collections are intended to be their own pages.

Whatever, added an extra directory and then an "if" check to all of our
loops — not beautiful but it appears to be working.

Ran all of the "play" titles through DeepL, so they're mostly
for demonstration.

3. Added langauge switching links

Super basic but they work okay.